Line Configuration

Line elements are used to represent the line in a line chart.

cjs_config_line(
  tension = 0,
  backgroundColor = NULL,
  borderWidth = 3,
  borderColor = NULL,
  borderCapStyle = "butt",
  borderDash = NULL,
  borderDashOffset = NULL,
  borderJoinStyle = "miter",
  capBezierPoints = TRUE,
  cubicInterpolationMode = "default",
  fill = FALSE,
  stepped = FALSE
)

Arguments

tension

number 0 Bézier curve tension (0 for no Bézier curves).

backgroundColor

Color Chart.defaults.backgroundColor Line fill color.

borderWidth

number 3 Line stroke width.

borderColor

Color Chart.defaults.borderColor Line stroke color.

borderCapStyle

string 'butt' Line cap style. See MDN

borderDash

number[] [] Line dash. See MDN

borderDashOffset

number 0.0 Line dash offset. See MDN

borderJoinStyle

'round'|'bevel'|'miter' 'miter' Line join style. See MDN

capBezierPoints

boolean true true to keep Bézier control inside the chart, false for no restriction.

cubicInterpolationMode

string 'default' Interpolation mode to apply. See more...

fill

boolean|string false How to fill the area under the line. See area charts.

stepped

boolean false true to show the line as a stepped line (tension will be ignored).

Value

a list