Edit the scale of the plot

Common options to all axes. Namespace: options.scales|scaleId| Note: depending on the axis type used, there will be more options. See cjs_scale_* functions. border, grid and ticks are themselves lists of options

cjs_scale_axis(
  p,
  id,
  type = NULL,
  alignToPixels = FALSE,
  backgroundColor = NULL,
  border = cjs_border(),
  display = TRUE,
  grid = cjs_grid(),
  min = NULL,
  max = NULL,
  reverse = FALSE,
  stacked = FALSE,
  suggestedMax = NULL,
  suggestedMin = NULL,
  ticks = cjs_ticks(),
  weight = 0
)

Arguments

p

plot

id

axis id

type

string Type of scale being employed. Custom scales can be created and registered with a string key. This allows changing the type of an axis for a chart.

alignToPixels

boolean FALSE Align pixel values to device pixels.

backgroundColor

Color Background color of the scale area.

border

object Border configuration. See cjs_border()

display

boolean|string TRUE Controls the axis global visibility (visible when true, hidden when false). When display: 'auto', the axis is visible only if at least one associated dataset is visible.

grid

object Grid line configuration. See cjs_grid()

min

number User defined minimum number for the scale, overrides minimum value from data. more...

max

number User defined maximum number for the scale, overrides maximum value from data. more...

reverse

boolean FALSE Reverse the scale.

stacked

boolean|string FALSE Should the data be stacked. more...

suggestedMax

number Adjustment used when calculating the maximum data value. more...

suggestedMin

number Adjustment used when calculating the minimum data value. more...

ticks

object Tick configuration. See cjs_ticks()

weight

number 0 The weight used to sort the axis. Higher weights are further away from the chart area.

Value

An object of class htmlwidget